Perbaiki pemilik yang ditolak dari datanya sendiri
Membuka detail pesanan sendiri di server berakhir dengan "Pesanan tersebut
bukan milik akun Anda", padahal daftarnya — yang dibatasi relasi milik
pengguna — menampilkan pesanan itu sebagai miliknya. Dua pemeriksaan yang
seharusnya sepakat justru berlawanan.

Penyebabnya tipe data, bukan kepemilikan. PDO dengan emulated prepares, lazim
pada hosting bersama, mengembalikan kolom angka sebagai string. Kunci asing
user_id tidak punya cast, sehingga terbaca "12" sementara auth()->id() bernilai
12; perbandingan ketat !== selalu benar dan pemiliknya ditolak. Di komputer
pengembang driver mengembalikan integer, jadi gejalanya tidak pernah muncul
selama pengembangan maupun pengujian.

Kunci asing pada Pesanan, Alamat, dan Keranjang kini di-cast menjadi integer,
dan seluruh perbandingan kepemilikan dibuat tahan tipe sebagai lapis kedua.

Ditambah pemeriksaan di sistem:cek yang melaporkan apakah driver
mengembalikan angka sebagai string, sebab keadaan itu menjelaskan gejala yang
hanya tampak di server.

79 uji lulus, termasuk yang menirukan kunci asing terbaca sebagai string.

+++ app/Http/Controllers/AlamatController.php
@@ -114,6 +114,6 @@ class AlamatController extends Controller
 
     private function authorizeOwn(Alamat $alamat): void
     {
-        abort_if($alamat->user_id !== auth()->id(), 403);
+        abort_if((int) $alamat->user_id !== (int) auth()->id(), 403);
     }
 }
